+++ packages/dock-blockchain-api/src/utils/index.js
@@ -1,3 +1,0 @@
-export * from './block';
-export * from './extrinsic';
-export * from './chain-ops';