+++ scripts/migration/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @docknetwork/migration
 
+## 0.9.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @docknetwork/cheqd-blockchain-api@0.35.0
+  - @docknetwork/credential-sdk@0.49.0
+  - @docknetwork/dock-blockchain-modules@0.28.1
+
 ## 0.9.1
 
 ### Patch Changes